Never use set-header-data for fast-bindable bit
x86-64 did not use set-header-data but there was quite likely a data race
on all other 64-bit machines because writing the entire header word would
need to be mutually exclusive with ENSURE-SYMBOL-TLS-INDEX. Instead use
sap-ref-8. This was only an issue if TLS index is not in a slot of its own.