ppc64: Switch to #+linkage-space for full calls
The call sequence needs only 1 add and 1 load where it used to need 3
loads for fdefn, fun, raw-fun. So while it is potentially faster, the
motivation was actually to remove support for untagged-fdefns which
strew #+/- all over the place.

The regression suite on ppc64 seems to be no more unstable than is typical.
I haven't observed a clean run in probably a year, and it may also be
getting some memory-ordering-related bugs due to the hash-table patches.