Somewhat hide XSETs assumption about EQL-hash
There's not a way to directly regression test that XSET-ELT-HASH is stable
since EQL-HASH is only unstable with #+(and aslr relocatable-static-space)
which isn't a config tested among our myriad CI builds. Theoretically we could
make a shell test that starts SBCL 2 or 3 times, looking for any changes in the
hash of NIL, but there's a second problem which is that SBCL often crashed
at startup on the gcc farm machine I was using, and it doesn't usually.
So this config isn't reliable enough to test on anyway from what I see.