Bind *CLOSE-STREAMS-ON-ERROR* in RUN-PROGRAM
RUN-PROGRAM calls GET-DESCRIPTOR-FOR which pushes stream objects
onto *CLOSE-STREAMS-ON-ERROR*. RUN-PROGRAM later reads the value
of *CLOSE-STREAMS-ON-ERROR* and closes the streams.

Before this change, both accesses used the global value
of *CLOSE-STREAMS-ON-ERROR* so that stream objects accumulated there
over time.

This change makes RUN-PROGRAM bind *CLOSE-STREAMS-ON-ERROR* (like it
already did for *CLOSE-FDS-ON-ERROR*) so that the above operations
only affect the dynamic scope of the binding. This change also removes
the global value of *CLOSE-STREAMS-ON-ERROR* so that a similar mistake
will lead to an error in the future.

The problem was introduced in commit b6732ebc.